AskiMate_main_platform/ai_app/requirements.txt
fastapi
uvicorn[standard]
rasa
huggingface_hub
pyyaml
pydantic
requests            # (huggingface_hub dependency, but often useful directly)
aiohttp 
